lacking a box (read comment)
The bounding box has values like -2, 0 for X (and alike for other coordinates). This taken from the point cloud, so it should be in meters, correct it to pixels?? (i have not clue if that is what is needed, or we can always work on meters and point clouds)
The testing_env.m is incomplete for now.